requirements.txt
<<<<<<< HEAD
flask
scikit-learn
pandas
numpy
tensorflow-cpu
opencv-python-headless
=======
numpy==1.21.2
streamlit==0.87.0
keras==2.4.3
opencv-contrib-python-headless
tensorflow-cpu
streamlit_webrt
>>>>>>> faffa8d (updated commit)
